requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.13.0
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
neurokit2==0.2.7
python-dateutil==2.8.2
